refactor(live): extract calculation sanity layer
